Adding a dependency on model_variables.inc
mpas_grid_types needs to depend on model_variables.inc, in order to
ensure the proper namelist name gets used. In the case where none of the
Registry files are updated, but the build options change, the namelist
name could change requiring a different mpas_grid_types to be generated.