Updating pool routines and fixing bugs.
Adding an optional argument allows the time levels in srcPool to be
overwritten using an input time levels. This is useful if you want to
clone to a pool and have a single time level in the output pool.

When copying pools, only copy as many time levels as the destination field
has.

Adding an initialize time levels routine that can be used to copy time
level 1 into all subsequent time levels, to initialize time levels with
real data.

Adding subroutine to print members of pools.

Fixing some bugs with interfaces and get_array routines.