LI: Modify LI Makefile to build dycore interface
This requires C++ compiler support in the main MPAS Makefile.  Logic has been
added to only build the interface if one of the external dycores has been
enabled.  The building of the interface C++ code has been done in a general
template for C++ code that can be extended to additional C++ files.