Avoid returning a zero denominator when the numerator is zero in mpas_get_timeInterval().
Often, it can be useful to query a time interval as some integral number of seconds (S),
plus a numerator (S_n) and denominator (S_d) to represent the fractional seconds in
the interval.  The previous version of the stand-alone ESMF time manager code would not
return a zero value for S_d even if the value of S_n was zero; this was convenient, since
we could always compute S + S_n / S_d without needing to check the value of S_d.
With this commit, we return to this behavior by setting S_d to 1 if S_n is zero.