LI: Add external velo solver functionality
This commit adds the MPAS code needed to use external dycores with MPAS.
This covers the L1L2, FO, and Stokes solvers available in LifeV,
Albany, and PHG.

This is a merge squash of the old commits 27b5c1 through ba800d in the
landice/add_external_dycore_interface branch from the MPAS-Release repo
(which may soon be deleted).  They needed to be updated for Pools, and
it was much easier to do it this way than preserve that old history,
most of which was corrections, changing names of things, and minor
updates to the API.

This commit does *not* include the changes to the build system to
include C++ interface code - that will follow in a separate commit.

The external dycore wrapper is contained in the new
mpas_li_velocity_external.F module.  New "velocity_solver" options are
now available when compiling with external velocity solver libraries:
'L1L2', 'FO', 'Stokes'

Some details about this commit:

* Calculation of anyVertexMaskChanged: if the vertex mask does not
  change between timesteps, then the external solvers do not need to
  recalculate their mesh information.  MPAS now does a check for this
  considering the following to get the initial time setup correctly:
  1) The calculation of anyVertexMaskChanged in
    diagnostic_solve_before_velocity() has to be saved to the specified
    timeLevel instead of always getting saved into time level 2 (since the
    initial solve currently is using time level 1).
  2) vertexMask on time level 2 has to be initialized to impossible
    values so that the first call of diagnostic_solve_before_velocity() will
    think the mask has changed and therefore set anyVertexMaskChanged=1.

* The array_from_exchange_list routine has been created to convert MPAS
  exchange lists into flat arrays to be used by external C/C++ velocity
  solvers.  It makes use of dmpar functionality to convert an exchage list
  to a communication list and then converts the communication list to a
  specified flat array format.  External dycores do not support multiple
  blocks per processor, but this implementation is written to support them
  where convenient.

* External dycores need config_num_halos>=2.

* The interface allows lifev to handle spheric geometries

* The name of files, subroutines, flags and namelist options in this
  commit are meant to be generic (i.e., not specific to LifeV, Albany, or
  PHG).  Unlike the earliest implementation of the interface, this commit
  assumes the API is fairly generic and does not care what code is on the
  other side of the interface, as long as it works!  There is  logic in
  the Makefile and in ifdefs in the code to only allow certain external
  libraries/solvers to do the things they are currently able to do.  This
  will need to be updated as libraries evolve (e.g., if Albany becomes
  able to do L1L2 or Stokes solves).
    Details:
    * To build with an external library, you must specify on the 'make'
    line one or more of: ALBANY=true, LIFEV=true, PHG=true
    * There is logic in the Makefile that defines new variables based
    on the external library being linked (the flags above):
    USE_EXTERNAL_L1L2, USE_EXTERNAL_FIRSTORDER, USE_EXTERNAL_STOKES
    * ifdefs in the Fortran code match these Makefile variable names
    * subroutines in mpas_li_velocity_external.F have the albany, lifev
    and phg names avoided where appropriate to have a more general name
    * namelist options for the velocity solvers have had the library
    names removed.  Valid options are now 'L1L2', 'FO', 'Stokes'.  So if you
    specify 'FO', you will use Albany if you've compiled with Albany or LifeV if you've
    compiled with LifeV - you don't have to change the namelist option based on which
    external library you compiled with.

I have tested this commit with standalone MPAS/sia and confirmed that
the code builds and the dome and EISMINT test cases remain unchanged.
I tried compiling with ALBANY=true but since this commit does not include
the C++ interface code that came from LifeV, I got linking errors. And I no
longer have a working copy of LIFEV conveniently available.