LI: Modify get_prism_velocity_on_FEdges for non-uniform mehses.
The function computes the average normal velocity on the edges/faces of
MPAS cells. It relies on the assumptions that
1. the locations of the circumcenters of two triangles sharing an edge are inside the union of the triangles
2. the locations of the cicumcenters should not coincide and should not be inverted
(i.e. if triangle 1 precedes triangle 2 on the circumcenters line, then circumcenter 1 should preced circumceter 2).
The flux obtained multiplying normal velocity with the area of the face
corresponding that edge will be exact for linear-bilinear finte elements on the prisms.
The flux will be second order accurate for other finite elements (e.g. linear or quadratic on Tetra).