LI: add cell center velos to ext. dycore interface
Previously, the external dycores were determining normalVelocity at
edges and just returning that, and then MPAS was using vector
reconstruction to calculate velocities at cell centers (uReconstructX,
etc.).  This was silly because the external dycores are natively solving
at cell center locations so we can just get those values directly.

This commit adds the cell center velocity fields to the L1L2, FO, and
Stokes interface calls for solving velocity.  It does NOT modify the
interface itself, so the code will not run with external dycores.