LI: Create package for HO variables
Model variables that are only needed for higher-order dycores are only
allocated and only read from input file if a higher-order dycore is
selected in the namelist file.

There currently is an issue in framework (#296) with streams sharing the
same filename_template.  So for now the inputHigherOrderVelocity stream
has a filename_template of landice_grid.nc2.  This needs to be changed
to landice_grid.nc in the streams.landice file before running the model.