Modified the broadcast routines and added maxloc/minloc routines in mpas_dmpar.F
First, I modified the broadcast subroutines to support some new landice diagnostics.
The following routines have been modified:
    mpas_dmpar_bcast_int
    mpas_dmpar_bcast_ints
    mpas_dmpar_bcast_real
    mpas_dmpar_bcast_reals
    mpas_dmpar_bcast_double
    mpas_dmpar_bcast_doubles
    mpas_dmpar_bcast_logical
    mpas_dmpar_bcast_char

In the old versions, it was assumed that IO_NODE is the source code
for the broadcast. In the modified versions, the user can specify the
source node by passing in an optional argument 'proc'. If no optional
argument is supplied, then IO_NODE is the source node. Thus existing
calls to these routines have the same behavior as before and need not
be modified.

I verified that mpas_dmpar_bcast_int and mpas_dmpar_bcast_real are
working for the new landice statistics module. The changes in the other
routines are basically identical, but I have not tested each one.

Second, I added the following four new subroutines:
    mpas_dmpar_minloc_int
    mpas_dmpar_minloc_real
    mpas_dmpar_maxloc_int
    mpas_dmpar_maxloc_real

These are similar to the existing routines mpas_dmpar_max_int, etc.
Like the existing routines, the new routines return a global max/min value.
The difference is that the new routines also return the processor on
which this max/min value resides. It may be desirable to have the processor
number in order to broadcast certain information to the head processor.
(For the landice statistics module, we broadcast the cell/edge global indices
associated with, e.g., the maximum ice thickness or velocity, and to do this
we need the processor number.)

I verified that the new maxloc/minloc routines are working as expected.
Similar routines are used in CISM.