Correct bug in paraview adapter.
There is a connectivity check of verticesOnCell in the adapter.
Previously, this used verticesPerCell, which is maxEdges.
It needs to use nEdgesOnCell[i], so it only checks the actual
vertices.  This error was revealed going from v2 to v3 because
in V3, verticesOnCell that don't exist (e.g. higher than index 6
on a hex) now correctly point to nVertices+1.  On V2 they
incorrectly pointed to a smaller number.