Adding the ability to specify a pool size in registry
This commit adds a new attribute to the specification of a var_struct.
Any var_struct can now have the pool_size attribute, which will specify
the size of the pool the var_struct will generate. By default the size
will be MPAS_POOL_TABLE_SIZE, but the size can be set to 1 causing the
pool to become a linked list ordered based on the order items are added
to the pool.