Switching the jm EOS to use zMid instead of refBottomDepth
For configurations where ssh differs significantly from zero,
refBottomDepth is not a good approximation of the depth of a
given layer.

I therefore changed the jm EOS to use zMid to compute the
reference pressure.  This change requires that zMid be computed
in diagnostics *before* the calls to EOS and regardless of how
the pressure gradient is computed.

zMid must also be passed to the jm EOS, as jm doesn't have access
to the diagnosticsPool.