Calculate normalSlopeEdge if DCFL check is enabled w/HO dycore
This adds normalSlopeEdge variable to calcDiffusivity package and
calculates it if either velocity solver is SIA *or* the DCFL check
is enabled.

Without this fix if the adaptive time stepper is set to calculate
diffusivity when a HO dycore is used, there is a segfault because
normalSlopeEdge is not allocated or calculated.