Add betaSolve variable to be used by velo solver
This commit makes it so there are two beta fields in the model when a HO
dycore is being used.  'beta' is still the field that is input by the
user.  'betaSolve' is a copy of 'beta' that is updated on each time step
to be 0 where the ice is floating.  This allows for grounding line
movement while retaining a beta field that can cover the entire domain.