Add in basal mass balance to geometry evolution
There are two components to basal mass balance:
groundedBasalMassBal - to be calculated internally by heat balance once
it is added
floatingBasalMassBal - input to the model from file or climate model

These two fields are combined on each time step into the field
basalMassBal which is what is actually used during evolution.  Therefore
basalMassBal is a useful output field but it is not an input field.