Add packages to HO input/restart variables
This way those fields will only be attempted to be read if a HO dycore
is being used.

This commit also adds dirichletVelocityMask as a restart field so that
time-evolving runs with Dirichlet b.c. will work correctly (e.g.
MISMIP).  Note that this means that this extra field will be present in
all HO restart files even if they don't use Dirichlet b.c.