Adding a tool to parse filter information
This commit adds a new tool within the context of registry parsing to
extract filter information from an existing registry file. It can be
used to generate filter_function.inc, which can be used to define the
filter's for a core.

It parses filter and filter_list blocks from a registry file, with an
example as follows:

    <filters>
        <filter name="test1" module_name="mpas_test_filter1"/>
        <filter name="test2" module_name="mpas_test_filter2"/>
        <filter name="test3" module_name="mpas_test_filter1"/>
    </filters>
    <filters>
        <filter name="test4" module_name="mpas_test_filter2"/>
        <filter name="test5" module_name="mpas_test_filter1"/>
        <filter name="test6" module_name="mpas_test_filter2"/>
    </filters>

    <filter_list name="group1">
        <filter name="test4"/>
        <filter name="test3"/>
        <filter name="test1"/>
    </filter_list>

    <filter_list name="group2">
        <filter name="test5"/>
        <filter name="test2"/>
        <filter name="test6"/>
    </filter_list>