Adding filters to the build system
This commit adds filters to the build system in a shared src/filters
directory. Additionally, this commit adds a driver which may be used to
act on lists of filters. It is provided to allow cores to utilize shared
functionality of a driver, without writing their own custom driver, but
is not required to use filters.

This commit also adds an example filter as
mpas_filter_test_compute_interval. This filter can be used to test that
the compute interval portion of the driver works correctly.