Add spatially variable shortwave radiation penetration
This commit will introduce variation of penetrating shortwave radiation
that is dependent on observed chlorophyll concentrations (satellite).
The actual parameterization will be based on Manizza et al. (2005),
Bio-optical feedbacks among phytoplankton, upper ocean physics and
sea-ice in a global model. Geophys. Res. Let. 32, L05603,
doi:10.1029/2004GL020778

Some of the hooks are in, and the registry has the necessary arrays.