change init test case from soma to baroclinic channel.
initially i planned to use soma as the test case for this, but had
trouble getting soma to work, even without these changes.  so i opted to
use a 'supported' case:  the baroclinic channel. the code mods allow the
ability to choose which cell to add runoff to as well as the magnitude
of runoff, precip, and evap.