Move cvmix vertViscTopOfCell to ocn_vmix_coefs
When CVMix is used, vertViscTopOfEdge needs to be calculated by
averaging vertViscTopOfCell to edges.
This averaging is performed in ocn_vmix_implicit after ocn_vmix_coefs is
called.
Before this averaging is done, vertViscTopOfEdge is reset to zero,
effectively reseting it.
In order to be able to increment vertViscTopOfEdge with other
parameterizations, this commit moves the averaging to ocn_vmix_coefs,
after ocn_vmix_coefs_cvmix_build is called.
This way, increments to vertViscTopOfEdge must be placed after this
averaging, not before, so they don't get reset.