Regional stats analysis member.
This analysis member performs regional statistics over masked areas,
such that it requires an external mask stream that defines the masks.
The mask stream input requires an (regions, cells) 2D array of integers,
with (0 or 1s) indicating if mesh elements are in or out of a region.
It also requires that it contains a list of groupings of regions.

Namelist configuration picks a group of region masks out of the stream,
while stream configuration describes the mask input and the input/output
stream of the fields that regional statistics will be computed over.
The user can pick avg, min, or max over all of the fields, and all
regions will be subsetted for all fields.