Introduce a new 3-d divergence damping scheme to damp acoustic modes
This commit replaces the old divergence damping mechanisms, including
the forward extrapolation of pressure, and removes its associated namelist
variable config_smdiv_p_forward and the divergence_3d field. The new mechanism
does not require any scaling of the smdiv coefficient on variable-resolution
meshes. Divergence damping is applied after the completion of each acoustic
step, immediately after the halo for rtheta_p is communicated; in this sense,
it is an addition to the acoustic step.