Threaded version of physics_get_tend in mpas_atmphys_todynamics.F, this changes the interface and the call to it in mpas_atm_time_integration.F. This also includes a threaded version of tend_toEdges in mpas_atm_todynamics.F. In order to keep the old, unused physics_addtend routine in the code, a legacy interface to the new, threaded version of tend_toEdges is provided.
A new OpenMP implementation is used in atm_srk3, which creates the threads only once for the entire time integration step and uses them on demand; non-threaded sections are enclosed in master-only blocks. Additional timers in atm_srk3 are added for most of the routines called to provide a finer granularity when profiling.