Removed if-test around call to update_radiation_diagnostics in mpas_atmphys_driver.F
Previously, the routine update_radiation_diagnostics to update the accumulated
radiation variables (acswdnb etc) was only called if config_bucket_update /= 'none'
and config_bucket_radt .gt. 0._RKIND. Inside the routine, the current values of
the (temporary) fields (swdnb etc) are multiplied by dt_dyn and added to the
accumulated (permanent) fields. Only after that, another test of config_bucket_update
and config_bucket_radt is done and the overflow values transferred to the buckets,
if requested. If config_bucket_update is set to none, the information stored in
the temporary fields (swdnb etc) is lost at the end of the time step and the
accumulated fields remain zero for the entire integration. This is not the intended
method, as users may wish to get the information on accumulated radiation also
for shorter runs, for which the buckets may not be needed.